Renamed `settings.py` to `spider_settings.py`
