+++ src/tiny_web_crawler/__init__.py
@@ -1,2 +1,2 @@
 from tiny_web_crawler.core.spider import Spider
-from tiny_web_crawler.core.settings import SpiderSettings
+from tiny_web_crawler.core.spider_settings import SpiderSettings
