+++ src/tiny_web_crawler/core/spider_settings.py
